# swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "WHS-TodoList API",
    "version": "1.0",
    "description": "API specification for the WHS-TodoList application. Manages users, todos, and holidays.",
    "contact": {
      "name": "WHS-TodoList Team",
      "email": "contact@whs-todolist.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3001/api",
      "description": "Development server"
    },
    {
      "url": "https://your-domain.vercel.app/api",
      "description": "Production server"
    }
  ],
...
  "paths": {
...
    "/todos": {
      "get": {
        "tags": ["Todos"],
        "summary": "Get list of todos",
        "operationId": "getTodos",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["active", "completed", "deleted"]
            }
          },
          { "name": "search", "in": "query", "schema": { "type": "string" } },
          {
            "name": "sortBy",
            "in": "query",
            "schema": { "type": "string", "enum": ["dueDate", "createdAt"] }
          },
          {
            "name": "order",
            "in": "query",
            "schema": { "type": "string", "enum": ["asc", "desc"] }
          }
        ],
        "responses": {
          "200": {
            "description": "A list of todos",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Todo" }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Todos"],
        "summary": "Create a new todo",
        "operationId": "createTodo",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateTodoRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Todo created successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TodoResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" }
        }
      }
    },
...
    "/trash": {
      "get": {
        "tags": ["Trash"],
        "summary": "Get items from trash",
        "operationId": "getTrashItems",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "schema": { "type": "string" },
            "description": "Search by title or content"
          },
          {
            "name": "sortBy",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["deletedAt", "dueDate", "createdAt"]
            },
            "description": "Sort by field"
          },
          {
            "name": "order",
            "in": "query",
            "schema": { "type": "string", "enum": ["asc", "desc"] },
            "description": "Sort order"
          }
        ],
        "responses": {
          "200": {
            "description": "A list of deleted todos",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Todo" }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/holidays": {
      "get": {
        "tags": ["Holidays"],
        "summary": "Get list of holidays",
        "operationId": "getHolidays",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "year", "in": "query", "schema": { "type": "integer" } },
          { "name": "month", "in": "query", "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": {
            "description": "A list of holidays",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Holiday" }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Holidays"],
        "summary": "Add a new holiday (Admin only)",
        "operationId": "createHoliday",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateHolidayRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Holiday created successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HolidayResponse" }
              }
            }
          },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
    "/holidays/{id}": {
      "put": {
        "tags": ["Holidays"],
        "summary": "Update a holiday (Admin only)",
        "operationId": "updateHoliday",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateHolidayRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Holiday updated successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HolidayResponse" }
              }
            }
          },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Todo": {
        "type": "object",
        "properties": {
          "todoId": { "type": "string", "format": "uuid" },
          "userId": { "type": "string", "format": "uuid" },
          "title": { "type": "string" },
          "content": { "type": "string", "nullable": true },
          "startDate": { "type": "string", "format": "date", "nullable": true },
          "dueDate": { "type": "string", "format": "date", "nullable": true },
          "status": {
            "type": "string",
            "enum": ["active", "completed", "deleted"]
          },
          "isCompleted": { "type": "boolean" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" },
          "deletedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
      "Holiday": {
        "type": "object",
        "properties": {
          "holidayId": { "type": "string", "format": "uuid" },
          "title": { "type": "string" },
          "date": { "type": "string", "format": "date" },
          "description": { "type": "string", "nullable": true },
          "isRecurring": { "type": "boolean" }
        }
      },
...
      "CreateTodoRequest": {
        "type": "object",
        "required": ["title"],
        "properties": {
          "title": { "type": "string", "example": "프로젝트 마감" },
          "content": { "type": "string", "example": "PRD 작성 완료하기" },
          "startDate": {
            "type": "string",
            "format": "date",
            "example": "2025-11-25"
          },
          "dueDate": {
            "type": "string",
            "format": "date",
            "example": "2025-11-28"
          }
        }
      },
...
      "CreateHolidayRequest": {
        "type": "object",
        "required": ["title", "date"],
        "properties": {
          "title": { "type": "string", "example": "광복절" },
          "date": {
            "type": "string",
            "format": "date",
            "example": "2025-08-15"
          },
          "description": {
            "type": "string",
            "example": "대한민국 독립 기념일"
          },
          "isRecurring": { "type": "boolean", "example": true }
        }
      },
      "UpdateHolidayRequest": {
        "type": "object",
        "properties": {
          "title": { "type": "string" },
          "date": { "type": "string", "format": "date" },
          "description": { "type": "string" },
          "isRecurring": { "type": "boolean" }
        }
      },
...
      "TodoResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": true },
          "data": { "$ref": "#/components/schemas/Todo" }
        }
      },
...
      "HolidayResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": true },
          "data": { "$ref": "#/components/schemas/Holiday" }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": false },
          "error": {
            "type": "object",
            "properties": {
              "code": { "type": "string" },
              "message": { "type": "string" }
            }
          }
        }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Bad Request - Invalid input",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      },
...
      "Forbidden": {
        "description": "Forbidden - Insufficient permissions",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      },
...
    }
  }
}